Handle rotation of the wallpaper/home screen.
Previously it would wipe the backing bitmap every time the home screen
changed orientation (because this triggers the resize callback). Now
I've changed it so the bitmap is square, and if the resize doesn't
increase the dimensions, it doesn't wipe it. This also entails adding an
extra transform in BubbleChamber.draw() to keep the bitmap centred on
screen.