Complete palette support
I've changed how it picks muon/antimuon pairs so that the colours are
paired in the palette ("0xff0000,0x00ff00") instead of it picking from
opposite ends of the list, to make it easier to write palettes. I've
also added some new palettes, and made it clear the bitmap and reset all
particles whenever you select a new palette.