Add About screen & feedback button
The feedback button is accomplished with a VIEW intent on a preference,
pointing to a mailto: URL. I tried using a SEND intent but it didn't
work.

There is one problem, which is that the about text changes to black when
you tap it. The links in the about text work fine, though.