Adjust particle parameters.
Base the number of particles on screen area, not length, so it can be
balanced correctly on larger screens.

Turn down the fading a bit.

Spread out quark velocities a little more and make the decay Gaussian.

Move the light and dark paths of hadrons closer together so they look
better on larger screens.