Slide side-to-side when the user changes home screens
Obey the "desired size" for the virtual size of the wallpaper, and use
onOffsetsChanged to pan. This lets it behave like wallpapers are
supposed to. It also entails moving around a lot of the min/max
dimension logic. I've made that more explicit by having a 'radius' that
the WallpaperEngine sets on BubbleChamber from the actual surface size.
This could be made into an option.